+++ .eslintignore
@@ -23,3 +23,4 @@ api/
 **/node_modules
 0*
 esm-resolver.js
+src/ext/diff/providers/*
