--- demo/diff/vscode-diff-legacy/index.js
@@ -1,2287 +1,0 @@
-'use strict';
-
-class DiffChange {
-  /**
-   * Constructs a new DiffChange with the given sequence information
-   * and content.
-   */
-  constructor(originalStart, originalLength, modifiedStart, modifiedLength) {
-    this.originalStart = originalStart;
-    this.originalLength = originalLength;
-    this.modifiedStart = modifiedStart;
-    this.modifiedLength = modifiedLength;
-  }
-  /**
-   * The end point (exclusive) of the change in the original sequence.
-   */
-  getOriginalEnd() {
-    return this.originalStart + this.originalLength;
-  }
-  /**
-   * The end point (exclusive) of the change in the modified sequence.
-   */
-  getModifiedEnd() {
-    return this.modifiedStart + this.modifiedLength;
-  }
-}
-
-var Constants = /* @__PURE__ */ ((Constants2) => {
-  Constants2[Constants2["MAX_SAFE_SMALL_INTEGER"] = 1073741824] = "MAX_SAFE_SMALL_INTEGER";
-  Constants2[Constants2["MIN_SAFE_SMALL_INTEGER"] = -1073741824] = "MIN_SAFE_SMALL_INTEGER";
-  Constants2[Constants2["MAX_UINT_8"] = 255] = "MAX_UINT_8";
-  Constants2[Constants2["MAX_UINT_16"] = 65535] = "MAX_UINT_16";
-  Constants2[Constants2["MAX_UINT_32"] = 4294967295] = "MAX_UINT_32";
-  Constants2[Constants2["UNICODE_SUPPLEMENTARY_PLANE_BEGIN"] = 65536] = "UNICODE_SUPPLEMENTARY_PLANE_BEGIN";
-  return Constants2;
-})(Constants || {});
-
-function firstNonWhitespaceIndex(str) {
-  for (let i = 0, len = str.length; i < len; i++) {
-    const chCode = str.charCodeAt(i);
-    if (chCode !== 32 && chCode !== 9) {
-      return i;
-    }
-  }
-  return -1;
-}
-function lastNonWhitespaceIndex(str, startIndex = str.length - 1) {
-  for (let i = startIndex; i >= 0; i--) {
-    const chCode = str.charCodeAt(i);
-    if (chCode !== 32 && chCode !== 9) {
-      return i;
-    }
-  }
-  return -1;
-}
-function isHighSurrogate(charCode) {
-  return 55296 <= charCode && charCode <= 56319;
-}
-function isLowSurrogate(charCode) {
-  return 56320 <= charCode && charCode <= 57343;
-}
-function computeCodePoint(highSurrogate, lowSurrogate) {
-  return (highSurrogate - 55296 << 10) + (lowSurrogate - 56320) + 65536;
-}
-
-function numberHash(val, initialHashVal) {
-  return (initialHashVal << 5) - initialHashVal + val | 0;
-}
-function stringHash(s, hashVal) {
-  hashVal = numberHash(149417, hashVal);
-  for (let i = 0, length = s.length; i < length; i++) {
-    hashVal = numberHash(s.charCodeAt(i), hashVal);
-  }
-  return hashVal;
-}
-function leftRotate(value, bits, totalBits = 32) {
-  const delta = totalBits - bits;
-  const mask = ~((1 << delta) - 1);
-  return (value << bits | (mask & value) >>> delta) >>> 0;
-}
-function toHexString(bufferOrValue, bitsize = 32) {
-  if (bufferOrValue instanceof ArrayBuffer) {
-    return Array.from(new Uint8Array(bufferOrValue)).map((b) => b.toString(16).padStart(2, "0")).join("");
-  }
-  return (bufferOrValue >>> 0).toString(16).padStart(bitsize / 4, "0");
-}
-class StringSHA1 {
-  constructor() {
-    // 80 * 4 = 320
-    this._h0 = 1732584193;
-    this._h1 = 4023233417;
-    this._h2 = 2562383102;
-    this._h3 = 271733878;
-    this._h4 = 3285377520;
-    this._buff = new Uint8Array(
-      64 /* BLOCK_SIZE */ + 3
-      /* to fit any utf-8 */
-    );
-    this._buffDV = new DataView(this._buff.buffer);
-    this._buffLen = 0;
-    this._totalLen = 0;
-    this._leftoverHighSurrogate = 0;
-    this._finished = false;
-  }
-  static {
-    this._bigBlock32 = new DataView(new ArrayBuffer(320));
-  }
-  update(str) {
-    const strLen = str.length;
-    if (strLen === 0) {
-      return;
-    }
-    const buff = this._buff;
-    let buffLen = this._buffLen;
-    let leftoverHighSurrogate = this._leftoverHighSurrogate;
-    let charCode;
-    let offset;
-    if (leftoverHighSurrogate !== 0) {
-      charCode = leftoverHighSurrogate;
-      offset = -1;
-      leftoverHighSurrogate = 0;
-    } else {
-      charCode = str.charCodeAt(0);
-      offset = 0;
-    }
-    while (true) {
-      let codePoint = charCode;
-      if (isHighSurrogate(charCode)) {
-        if (offset + 1 < strLen) {
-          const nextCharCode = str.charCodeAt(offset + 1);
-          if (isLowSurrogate(nextCharCode)) {
-            offset++;
-            codePoint = computeCodePoint(charCode, nextCharCode);
-          } else {
-            codePoint = 65533 /* UNICODE_REPLACEMENT */;
-          }
-        } else {
-          leftoverHighSurrogate = charCode;
-          break;
-        }
-      } else if (isLowSurrogate(charCode)) {
-        codePoint = 65533 /* UNICODE_REPLACEMENT */;
-      }
-      buffLen = this._push(buff, buffLen, codePoint);
-      offset++;
-      if (offset < strLen) {
-        charCode = str.charCodeAt(offset);
-      } else {
-        break;
-      }
-    }
-    this._buffLen = buffLen;
-    this._leftoverHighSurrogate = leftoverHighSurrogate;
-  }
-  _push(buff, buffLen, codePoint) {
-    if (codePoint < 128) {
-      buff[buffLen++] = codePoint;
-    } else if (codePoint < 2048) {
-      buff[buffLen++] = 192 | (codePoint & 1984) >>> 6;
-      buff[buffLen++] = 128 | (codePoint & 63) >>> 0;
-    } else if (codePoint < 65536) {
-      buff[buffLen++] = 224 | (codePoint & 61440) >>> 12;
-      buff[buffLen++] = 128 | (codePoint & 4032) >>> 6;
-      buff[buffLen++] = 128 | (codePoint & 63) >>> 0;
-    } else {
-      buff[buffLen++] = 240 | (codePoint & 1835008) >>> 18;
-      buff[buffLen++] = 128 | (codePoint & 258048) >>> 12;
-      buff[buffLen++] = 128 | (codePoint & 4032) >>> 6;
-      buff[buffLen++] = 128 | (codePoint & 63) >>> 0;
-    }
-    if (buffLen >= 64 /* BLOCK_SIZE */) {
-      this._step();
-      buffLen -= 64 /* BLOCK_SIZE */;
-      this._totalLen += 64 /* BLOCK_SIZE */;
-      buff[0] = buff[64 /* BLOCK_SIZE */ + 0];
-      buff[1] = buff[64 /* BLOCK_SIZE */ + 1];
-      buff[2] = buff[64 /* BLOCK_SIZE */ + 2];
-    }
-    return buffLen;
-  }
-  digest() {
-    if (!this._finished) {
-      this._finished = true;
-      if (this._leftoverHighSurrogate) {
-        this._leftoverHighSurrogate = 0;
-        this._buffLen = this._push(this._buff, this._buffLen, 65533 /* UNICODE_REPLACEMENT */);
-      }
-      this._totalLen += this._buffLen;
-      this._wrapUp();
-    }
-    return toHexString(this._h0) + toHexString(this._h1) + toHexString(this._h2) + toHexString(this._h3) + toHexString(this._h4);
-  }
-  _wrapUp() {
-    this._buff[this._buffLen++] = 128;
-    this._buff.subarray(this._buffLen).fill(0);
-    if (this._buffLen > 56) {
-      this._step();
-      this._buff.fill(0);
-    }
-    const ml = 8 * this._totalLen;
-    this._buffDV.setUint32(56, Math.floor(ml / 4294967296), false);
-    this._buffDV.setUint32(60, ml % 4294967296, false);
-    this._step();
-  }
-  _step() {
-    const bigBlock32 = StringSHA1._bigBlock32;
-    const data = this._buffDV;
-    for (let j = 0; j < 64; j += 4) {
-      bigBlock32.setUint32(j, data.getUint32(j, false), false);
-    }
-    for (let j = 64; j < 320; j += 4) {
-      bigBlock32.setUint32(j, leftRotate(bigBlock32.getUint32(j - 12, false) ^ bigBlock32.getUint32(j - 32, false) ^ bigBlock32.getUint32(j - 56, false) ^ bigBlock32.getUint32(j - 64, false), 1), false);
-    }
-    let a = this._h0;
-    let b = this._h1;
-    let c = this._h2;
-    let d = this._h3;
-    let e = this._h4;
-    let f, k;
-    let temp;
-    for (let j = 0; j < 80; j++) {
-      if (j < 20) {
-        f = b & c | ~b & d;
-        k = 1518500249;
-      } else if (j < 40) {
-        f = b ^ c ^ d;
-        k = 1859775393;
-      } else if (j < 60) {
-        f = b & c | b & d | c & d;
-        k = 2400959708;
-      } else {
-        f = b ^ c ^ d;
-        k = 3395469782;
-      }
-      temp = leftRotate(a, 5) + f + e + k + bigBlock32.getUint32(j * 4, false) & 4294967295;
-      e = d;
-      d = c;
-      c = leftRotate(b, 30);
-      b = a;
-      a = temp;
-    }
-    this._h0 = this._h0 + a & 4294967295;
-    this._h1 = this._h1 + b & 4294967295;
-    this._h2 = this._h2 + c & 4294967295;
-    this._h3 = this._h3 + d & 4294967295;
-    this._h4 = this._h4 + e & 4294967295;
-  }
-}
-
-class Debug {
-  static Assert(condition, message) {
-    if (!condition) {
-      throw new Error(message);
-    }
-  }
-}
-class MyArray {
-  /**
-   * Copies a range of elements from an Array starting at the specified source index and pastes
-   * them to another Array starting at the specified destination index. The length and the indexes
-   * are specified as 64-bit integers.
-   * sourceArray:
-   *		The Array that contains the data to copy.
-   * sourceIndex:
-   *		A 64-bit integer that represents the index in the sourceArray at which copying begins.
-   * destinationArray:
-   *		The Array that receives the data.
-   * destinationIndex:
-   *		A 64-bit integer that represents the index in the destinationArray at which storing begins.
-   * length:
-   *		A 64-bit integer that represents the number of elements to copy.
-   */
-  static Copy(sourceArray, sourceIndex, destinationArray, destinationIndex, length) {
-    for (let i = 0; i < length; i++) {
-      destinationArray[destinationIndex + i] = sourceArray[sourceIndex + i];
-    }
-  }
-  static Copy2(sourceArray, sourceIndex, destinationArray, destinationIndex, length) {
-    for (let i = 0; i < length; i++) {
-      destinationArray[destinationIndex + i] = sourceArray[sourceIndex + i];
-    }
-  }
-}
-class DiffChangeHelper {
-  /**
-   * Constructs a new DiffChangeHelper for the given DiffSequences.
-   */
-  constructor() {
-    this.m_changes = [];
-    this.m_originalStart = Constants.MAX_SAFE_SMALL_INTEGER;
-    this.m_modifiedStart = Constants.MAX_SAFE_SMALL_INTEGER;
-    this.m_originalCount = 0;
-    this.m_modifiedCount = 0;
-  }
-  /**
-   * Marks the beginning of the next change in the set of differences.
-   */
-  MarkNextChange() {
-    if (this.m_originalCount > 0 || this.m_modifiedCount > 0) {
-      this.m_changes.push(new DiffChange(
-        this.m_originalStart,
-        this.m_originalCount,
-        this.m_modifiedStart,
-        this.m_modifiedCount
-      ));
-    }
-    this.m_originalCount = 0;
-    this.m_modifiedCount = 0;
-    this.m_originalStart = Constants.MAX_SAFE_SMALL_INTEGER;
-    this.m_modifiedStart = Constants.MAX_SAFE_SMALL_INTEGER;
-  }
-  /**
-   * Adds the original element at the given position to the elements
-   * affected by the current change. The modified index gives context
-   * to the change position with respect to the original sequence.
-   * @param originalIndex The index of the original element to add.
-   * @param modifiedIndex The index of the modified element that provides corresponding position in the modified sequence.
-   */
-  AddOriginalElement(originalIndex, modifiedIndex) {
-    this.m_originalStart = Math.min(this.m_originalStart, originalIndex);
-    this.m_modifiedStart = Math.min(this.m_modifiedStart, modifiedIndex);
-    this.m_originalCount++;
-  }
-  /**
-   * Adds the modified element at the given position to the elements
-   * affected by the current change. The original index gives context
-   * to the change position with respect to the modified sequence.
-   * @param originalIndex The index of the original element that provides corresponding position in the original sequence.
-   * @param modifiedIndex The index of the modified element to add.
-   */
-  AddModifiedElement(originalIndex, modifiedIndex) {
-    this.m_originalStart = Math.min(this.m_originalStart, originalIndex);
-    this.m_modifiedStart = Math.min(this.m_modifiedStart, modifiedIndex);
-    this.m_modifiedCount++;
-  }
-  /**
-   * Retrieves all of the changes marked by the class.
-   */
-  getChanges() {
-    if (this.m_originalCount > 0 || this.m_modifiedCount > 0) {
-      this.MarkNextChange();
-    }
-    return this.m_changes;
-  }
-  /**
-   * Retrieves all of the changes marked by the class in the reverse order
-   */
-  getReverseChanges() {
-    if (this.m_originalCount > 0 || this.m_modifiedCount > 0) {
-      this.MarkNextChange();
-    }
-    this.m_changes.reverse();
-    return this.m_changes;
-  }
-}
-class LcsDiff {
-  /**
-   * Constructs the DiffFinder
-   */
-  constructor(originalSequence, modifiedSequence, continueProcessingPredicate = null) {
-    this.ContinueProcessingPredicate = continueProcessingPredicate;
-    this._originalSequence = originalSequence;
-    this._modifiedSequence = modifiedSequence;
-    const [originalStringElements, originalElementsOrHash, originalHasStrings] = LcsDiff._getElements(originalSequence);
-    const [modifiedStringElements, modifiedElementsOrHash, modifiedHasStrings] = LcsDiff._getElements(modifiedSequence);
-    this._hasStrings = originalHasStrings && modifiedHasStrings;
-    this._originalStringElements = originalStringElements;
-    this._originalElementsOrHash = originalElementsOrHash;
-    this._modifiedStringElements = modifiedStringElements;
-    this._modifiedElementsOrHash = modifiedElementsOrHash;
-    this.m_forwardHistory = [];
-    this.m_reverseHistory = [];
-  }
-  static _isStringArray(arr) {
-    return arr.length > 0 && typeof arr[0] === "string";
-  }
-  static _getElements(sequence) {
-    const elements = sequence.getElements();
-    if (LcsDiff._isStringArray(elements)) {
-      const hashes = new Int32Array(elements.length);
-      for (let i = 0, len = elements.length; i < len; i++) {
-        hashes[i] = stringHash(elements[i], 0);
-      }
-      return [elements, hashes, true];
-    }
-    if (elements instanceof Int32Array) {
-      return [[], elements, false];
-    }
-    return [[], new Int32Array(elements), false];
-  }
-  ElementsAreEqual(originalIndex, newIndex) {
-    if (this._originalElementsOrHash[originalIndex] !== this._modifiedElementsOrHash[newIndex]) {
-      return false;
-    }
-    return this._hasStrings ? this._originalStringElements[originalIndex] === this._modifiedStringElements[newIndex] : true;
-  }
-  ElementsAreStrictEqual(originalIndex, newIndex) {
-    if (!this.ElementsAreEqual(originalIndex, newIndex)) {
-      return false;
-    }
-    const originalElement = LcsDiff._getStrictElement(this._originalSequence, originalIndex);
-    const modifiedElement = LcsDiff._getStrictElement(this._modifiedSequence, newIndex);
-    return originalElement === modifiedElement;
-  }
-  static _getStrictElement(sequence, index) {
-    if (typeof sequence.getStrictElement === "function") {
-      return sequence.getStrictElement(index);
-    }
-    return null;
-  }
-  OriginalElementsAreEqual(index1, index2) {
-    if (this._originalElementsOrHash[index1] !== this._originalElementsOrHash[index2]) {
-      return false;
-    }
-    return this._hasStrings ? this._originalStringElements[index1] === this._originalStringElements[index2] : true;
-  }
-  ModifiedElementsAreEqual(index1, index2) {
-    if (this._modifiedElementsOrHash[index1] !== this._modifiedElementsOrHash[index2]) {
-      return false;
-    }
-    return this._hasStrings ? this._modifiedStringElements[index1] === this._modifiedStringElements[index2] : true;
-  }
-  ComputeDiff(pretty) {
-    return this._ComputeDiff(0, this._originalElementsOrHash.length - 1, 0, this._modifiedElementsOrHash.length - 1, pretty);
-  }
-  /**
-   * Computes the differences between the original and modified input
-   * sequences on the bounded range.
-   * @returns An array of the differences between the two input sequences.
-   */
-  _ComputeDiff(originalStart, originalEnd, modifiedStart, modifiedEnd, pretty) {
-    const quitEarlyArr = [false];
-    let changes = this.ComputeDiffRecursive(originalStart, originalEnd, modifiedStart, modifiedEnd, quitEarlyArr);
-    if (pretty) {
-      changes = this.PrettifyChanges(changes);
-    }
-    return {
-      quitEarly: quitEarlyArr[0],
-      changes
-    };
-  }
-  /**
-   * Private helper method which computes the differences on the bounded range
-   * recursively.
-   * @returns An array of the differences between the two input sequences.
-   */
-  ComputeDiffRecursive(originalStart, originalEnd, modifiedStart, modifiedEnd, quitEarlyArr) {
-    quitEarlyArr[0] = false;
-    while (originalStart <= originalEnd && modifiedStart <= modifiedEnd && this.ElementsAreEqual(originalStart, modifiedStart)) {
-      originalStart++;
-      modifiedStart++;
-    }
-    while (originalEnd >= originalStart && modifiedEnd >= modifiedStart && this.ElementsAreEqual(originalEnd, modifiedEnd)) {
-      originalEnd--;
-      modifiedEnd--;
-    }
-    if (originalStart > originalEnd || modifiedStart > modifiedEnd) {
-      let changes;
-      if (modifiedStart <= modifiedEnd) {
-        Debug.Assert(originalStart === originalEnd + 1, "originalStart should only be one more than originalEnd");
-        changes = [
-          new DiffChange(originalStart, 0, modifiedStart, modifiedEnd - modifiedStart + 1)
-        ];
-      } else if (originalStart <= originalEnd) {
-        Debug.Assert(modifiedStart === modifiedEnd + 1, "modifiedStart should only be one more than modifiedEnd");
-        changes = [
-          new DiffChange(originalStart, originalEnd - originalStart + 1, modifiedStart, 0)
-        ];
-      } else {
-        Debug.Assert(originalStart === originalEnd + 1, "originalStart should only be one more than originalEnd");
-        Debug.Assert(modifiedStart === modifiedEnd + 1, "modifiedStart should only be one more than modifiedEnd");
-        changes = [];
-      }
-      return changes;
-    }
-    const midOriginalArr = [0];
-    const midModifiedArr = [0];
-    const result = this.ComputeRecursionPoint(originalStart, originalEnd, modifiedStart, modifiedEnd, midOriginalArr, midModifiedArr, quitEarlyArr);
-    const midOriginal = midOriginalArr[0];
-    const midModified = midModifiedArr[0];
-    if (result !== null) {
-      return result;
-    } else if (!quitEarlyArr[0]) {
-      const leftChanges = this.ComputeDiffRecursive(originalStart, midOriginal, modifiedStart, midModified, quitEarlyArr);
-      let rightChanges = [];
-      if (!quitEarlyArr[0]) {
-        rightChanges = this.ComputeDiffRecursive(midOriginal + 1, originalEnd, midModified + 1, modifiedEnd, quitEarlyArr);
-      } else {
-        rightChanges = [
-          new DiffChange(midOriginal + 1, originalEnd - (midOriginal + 1) + 1, midModified + 1, modifiedEnd - (midModified + 1) + 1)
-        ];
-      }
-      return this.ConcatenateChanges(leftChanges, rightChanges);
-    }
-    return [
-      new DiffChange(originalStart, originalEnd - originalStart + 1, modifiedStart, modifiedEnd - modifiedStart + 1)
-    ];
-  }
-  WALKTRACE(diagonalForwardBase, diagonalForwardStart, diagonalForwardEnd, diagonalForwardOffset, diagonalReverseBase, diagonalReverseStart, diagonalReverseEnd, diagonalReverseOffset, forwardPoints, reversePoints, originalIndex, originalEnd, midOriginalArr, modifiedIndex, modifiedEnd, midModifiedArr, deltaIsEven, quitEarlyArr) {
-    let forwardChanges = null;
-    let reverseChanges = null;
-    let changeHelper = new DiffChangeHelper();
-    let diagonalMin = diagonalForwardStart;
-    let diagonalMax = diagonalForwardEnd;
-    let diagonalRelative = midOriginalArr[0] - midModifiedArr[0] - diagonalForwardOffset;
-    let lastOriginalIndex = Constants.MIN_SAFE_SMALL_INTEGER;
-    let historyIndex = this.m_forwardHistory.length - 1;
-    do {
-      const diagonal = diagonalRelative + diagonalForwardBase;
-      if (diagonal === diagonalMin || diagonal < diagonalMax && forwardPoints[diagonal - 1] < forwardPoints[diagonal + 1]) {
-        originalIndex = forwardPoints[diagonal + 1];
-        modifiedIndex = originalIndex - diagonalRelative - diagonalForwardOffset;
-        if (originalIndex < lastOriginalIndex) {
-          changeHelper.MarkNextChange();
-        }
-        lastOriginalIndex = originalIndex;
-        changeHelper.AddModifiedElement(originalIndex + 1, modifiedIndex);
-        diagonalRelative = diagonal + 1 - diagonalForwardBase;
-      } else {
-        originalIndex = forwardPoints[diagonal - 1] + 1;
-        modifiedIndex = originalIndex - diagonalRelative - diagonalForwardOffset;
-        if (originalIndex < lastOriginalIndex) {
-          changeHelper.MarkNextChange();
-        }
-        lastOriginalIndex = originalIndex - 1;
-        changeHelper.AddOriginalElement(originalIndex, modifiedIndex + 1);
-        diagonalRelative = diagonal - 1 - diagonalForwardBase;
-      }
-      if (historyIndex >= 0) {
-        forwardPoints = this.m_forwardHistory[historyIndex];
-        diagonalForwardBase = forwardPoints[0];
-        diagonalMin = 1;
-        diagonalMax = forwardPoints.length - 1;
-      }
-    } while (--historyIndex >= -1);
-    forwardChanges = changeHelper.getReverseChanges();
-    if (quitEarlyArr[0]) {
-      let originalStartPoint = midOriginalArr[0] + 1;
-      let modifiedStartPoint = midModifiedArr[0] + 1;
-      if (forwardChanges !== null && forwardChanges.length > 0) {
-        const lastForwardChange = forwardChanges[forwardChanges.length - 1];
-        originalStartPoint = Math.max(originalStartPoint, lastForwardChange.getOriginalEnd());
-        modifiedStartPoint = Math.max(modifiedStartPoint, lastForwardChange.getModifiedEnd());
-      }
-      reverseChanges = [
-        new DiffChange(
-          originalStartPoint,
-          originalEnd - originalStartPoint + 1,
-          modifiedStartPoint,
-          modifiedEnd - modifiedStartPoint + 1
-        )
-      ];
-    } else {
-      changeHelper = new DiffChangeHelper();
-      diagonalMin = diagonalReverseStart;
-      diagonalMax = diagonalReverseEnd;
-      diagonalRelative = midOriginalArr[0] - midModifiedArr[0] - diagonalReverseOffset;
-      lastOriginalIndex = Constants.MAX_SAFE_SMALL_INTEGER;
-      historyIndex = deltaIsEven ? this.m_reverseHistory.length - 1 : this.m_reverseHistory.length - 2;
-      do {
-        const diagonal = diagonalRelative + diagonalReverseBase;
-        if (diagonal === diagonalMin || diagonal < diagonalMax && reversePoints[diagonal - 1] >= reversePoints[diagonal + 1]) {
-          originalIndex = reversePoints[diagonal + 1] - 1;
-          modifiedIndex = originalIndex - diagonalRelative - diagonalReverseOffset;
-          if (originalIndex > lastOriginalIndex) {
-            changeHelper.MarkNextChange();
-          }
-          lastOriginalIndex = originalIndex + 1;
-          changeHelper.AddOriginalElement(originalIndex + 1, modifiedIndex + 1);
-          diagonalRelative = diagonal + 1 - diagonalReverseBase;
-        } else {
-          originalIndex = reversePoints[diagonal - 1];
-          modifiedIndex = originalIndex - diagonalRelative - diagonalReverseOffset;
-          if (originalIndex > lastOriginalIndex) {
-            changeHelper.MarkNextChange();
-          }
-          lastOriginalIndex = originalIndex;
-          changeHelper.AddModifiedElement(originalIndex + 1, modifiedIndex + 1);
-          diagonalRelative = diagonal - 1 - diagonalReverseBase;
-        }
-        if (historyIndex >= 0) {
-          reversePoints = this.m_reverseHistory[historyIndex];
-          diagonalReverseBase = reversePoints[0];
-          diagonalMin = 1;
-          diagonalMax = reversePoints.length - 1;
-        }
-      } while (--historyIndex >= -1);
-      reverseChanges = changeHelper.getChanges();
-    }
-    return this.ConcatenateChanges(forwardChanges, reverseChanges);
-  }
-  /**
-   * Given the range to compute the diff on, this method finds the point:
-   * (midOriginal, midModified)
-   * that exists in the middle of the LCS of the two sequences and
-   * is the point at which the LCS problem may be broken down recursively.
-   * This method will try to keep the LCS trace in memory. If the LCS recursion
-   * point is calculated and the full trace is available in memory, then this method
-   * will return the change list.
-   * @param originalStart The start bound of the original sequence range
-   * @param originalEnd The end bound of the original sequence range
-   * @param modifiedStart The start bound of the modified sequence range
-   * @param modifiedEnd The end bound of the modified sequence range
-   * @param midOriginal The middle point of the original sequence range
-   * @param midModified The middle point of the modified sequence range
-   * @returns The diff changes, if available, otherwise null
-   */
-  ComputeRecursionPoint(originalStart, originalEnd, modifiedStart, modifiedEnd, midOriginalArr, midModifiedArr, quitEarlyArr) {
-    let originalIndex = 0, modifiedIndex = 0;
-    let diagonalForwardStart = 0, diagonalForwardEnd = 0;
-    let diagonalReverseStart = 0, diagonalReverseEnd = 0;
-    originalStart--;
-    modifiedStart--;
-    midOriginalArr[0] = 0;
-    midModifiedArr[0] = 0;
-    this.m_forwardHistory = [];
-    this.m_reverseHistory = [];
-    const maxDifferences = originalEnd - originalStart + (modifiedEnd - modifiedStart);
-    const numDiagonals = maxDifferences + 1;
-    const forwardPoints = new Int32Array(numDiagonals);
-    const reversePoints = new Int32Array(numDiagonals);
-    const diagonalForwardBase = modifiedEnd - modifiedStart;
-    const diagonalReverseBase = originalEnd - originalStart;
-    const diagonalForwardOffset = originalStart - modifiedStart;
-    const diagonalReverseOffset = originalEnd - modifiedEnd;
-    const delta = diagonalReverseBase - diagonalForwardBase;
-    const deltaIsEven = delta % 2 === 0;
-    forwardPoints[diagonalForwardBase] = originalStart;
-    reversePoints[diagonalReverseBase] = originalEnd;
-    quitEarlyArr[0] = false;
-    for (let numDifferences = 1; numDifferences <= maxDifferences / 2 + 1; numDifferences++) {
-      let furthestOriginalIndex = 0;
-      let furthestModifiedIndex = 0;
-      diagonalForwardStart = this.ClipDiagonalBound(diagonalForwardBase - numDifferences, numDifferences, diagonalForwardBase, numDiagonals);
-      diagonalForwardEnd = this.ClipDiagonalBound(diagonalForwardBase + numDifferences, numDifferences, diagonalForwardBase, numDiagonals);
-      for (let diagonal = diagonalForwardStart; diagonal <= diagonalForwardEnd; diagonal += 2) {
-        if (diagonal === diagonalForwardStart || diagonal < diagonalForwardEnd && forwardPoints[diagonal - 1] < forwardPoints[diagonal + 1]) {
-          originalIndex = forwardPoints[diagonal + 1];
-        } else {
-          originalIndex = forwardPoints[diagonal - 1] + 1;
-        }
-        modifiedIndex = originalIndex - (diagonal - diagonalForwardBase) - diagonalForwardOffset;
-        const tempOriginalIndex = originalIndex;
-        while (originalIndex < originalEnd && modifiedIndex < modifiedEnd && this.ElementsAreEqual(originalIndex + 1, modifiedIndex + 1)) {
-          originalIndex++;
-          modifiedIndex++;
-        }
-        forwardPoints[diagonal] = originalIndex;
-        if (originalIndex + modifiedIndex > furthestOriginalIndex + furthestModifiedIndex) {
-          furthestOriginalIndex = originalIndex;
-          furthestModifiedIndex = modifiedIndex;
-        }
-        if (!deltaIsEven && Math.abs(diagonal - diagonalReverseBase) <= numDifferences - 1) {
-          if (originalIndex >= reversePoints[diagonal]) {
-            midOriginalArr[0] = originalIndex;
-            midModifiedArr[0] = modifiedIndex;
-            if (tempOriginalIndex <= reversePoints[diagonal] && 1447 /* MaxDifferencesHistory */ > 0 && numDifferences <= 1447 /* MaxDifferencesHistory */ + 1) {
-              return this.WALKTRACE(
-                diagonalForwardBase,
-                diagonalForwardStart,
-                diagonalForwardEnd,
-                diagonalForwardOffset,
-                diagonalReverseBase,
-                diagonalReverseStart,
-                diagonalReverseEnd,
-                diagonalReverseOffset,
-                forwardPoints,
-                reversePoints,
-                originalIndex,
-                originalEnd,
-                midOriginalArr,
-                modifiedIndex,
-                modifiedEnd,
-                midModifiedArr,
-                deltaIsEven,
-                quitEarlyArr
-              );
-            } else {
-              return null;
-            }
-          }
-        }
-      }
-      const matchLengthOfLongest = (furthestOriginalIndex - originalStart + (furthestModifiedIndex - modifiedStart) - numDifferences) / 2;
-      if (this.ContinueProcessingPredicate !== null && !this.ContinueProcessingPredicate(furthestOriginalIndex, matchLengthOfLongest)) {
-        quitEarlyArr[0] = true;
-        midOriginalArr[0] = furthestOriginalIndex;
-        midModifiedArr[0] = furthestModifiedIndex;
-        if (matchLengthOfLongest > 0 && 1447 /* MaxDifferencesHistory */ > 0 && numDifferences <= 1447 /* MaxDifferencesHistory */ + 1) {
-          return this.WALKTRACE(
-            diagonalForwardBase,
-            diagonalForwardStart,
-            diagonalForwardEnd,
-            diagonalForwardOffset,
-            diagonalReverseBase,
-            diagonalReverseStart,
-            diagonalReverseEnd,
-            diagonalReverseOffset,
-            forwardPoints,
-            reversePoints,
-            originalIndex,
-            originalEnd,
-            midOriginalArr,
-            modifiedIndex,
-            modifiedEnd,
-            midModifiedArr,
-            deltaIsEven,
-            quitEarlyArr
-          );
-        } else {
-          originalStart++;
-          modifiedStart++;
-          return [
-            new DiffChange(
-              originalStart,
-              originalEnd - originalStart + 1,
-              modifiedStart,
-              modifiedEnd - modifiedStart + 1
-            )
-          ];
-        }
-      }
-      diagonalReverseStart = this.ClipDiagonalBound(diagonalReverseBase - numDifferences, numDifferences, diagonalReverseBase, numDiagonals);
-      diagonalReverseEnd = this.ClipDiagonalBound(diagonalReverseBase + numDifferences, numDifferences, diagonalReverseBase, numDiagonals);
-      for (let diagonal = diagonalReverseStart; diagonal <= diagonalReverseEnd; diagonal += 2) {
-        if (diagonal === diagonalReverseStart || diagonal < diagonalReverseEnd && reversePoints[diagonal - 1] >= reversePoints[diagonal + 1]) {
-          originalIndex = reversePoints[diagonal + 1] - 1;
-        } else {
-          originalIndex = reversePoints[diagonal - 1];
-        }
-        modifiedIndex = originalIndex - (diagonal - diagonalReverseBase) - diagonalReverseOffset;
-        const tempOriginalIndex = originalIndex;
-        while (originalIndex > originalStart && modifiedIndex > modifiedStart && this.ElementsAreEqual(originalIndex, modifiedIndex)) {
-          originalIndex--;
-          modifiedIndex--;
-        }
-        reversePoints[diagonal] = originalIndex;
-        if (deltaIsEven && Math.abs(diagonal - diagonalForwardBase) <= numDifferences) {
-          if (originalIndex <= forwardPoints[diagonal]) {
-            midOriginalArr[0] = originalIndex;
-            midModifiedArr[0] = modifiedIndex;
-            if (tempOriginalIndex >= forwardPoints[diagonal] && 1447 /* MaxDifferencesHistory */ > 0 && numDifferences <= 1447 /* MaxDifferencesHistory */ + 1) {
-              return this.WALKTRACE(
-                diagonalForwardBase,
-                diagonalForwardStart,
-                diagonalForwardEnd,
-                diagonalForwardOffset,
-                diagonalReverseBase,
-                diagonalReverseStart,
-                diagonalReverseEnd,
-                diagonalReverseOffset,
-                forwardPoints,
-                reversePoints,
-                originalIndex,
-                originalEnd,
-                midOriginalArr,
-                modifiedIndex,
-                modifiedEnd,
-                midModifiedArr,
-                deltaIsEven,
-                quitEarlyArr
-              );
-            } else {
-              return null;
-            }
-          }
-        }
-      }
-      if (numDifferences <= 1447 /* MaxDifferencesHistory */) {
-        let temp = new Int32Array(diagonalForwardEnd - diagonalForwardStart + 2);
-        temp[0] = diagonalForwardBase - diagonalForwardStart + 1;
-        MyArray.Copy2(forwardPoints, diagonalForwardStart, temp, 1, diagonalForwardEnd - diagonalForwardStart + 1);
-        this.m_forwardHistory.push(temp);
-        temp = new Int32Array(diagonalReverseEnd - diagonalReverseStart + 2);
-        temp[0] = diagonalReverseBase - diagonalReverseStart + 1;
-        MyArray.Copy2(reversePoints, diagonalReverseStart, temp, 1, diagonalReverseEnd - diagonalReverseStart + 1);
-        this.m_reverseHistory.push(temp);
-      }
-    }
-    return this.WALKTRACE(
-      diagonalForwardBase,
-      diagonalForwardStart,
-      diagonalForwardEnd,
-      diagonalForwardOffset,
-      diagonalReverseBase,
-      diagonalReverseStart,
-      diagonalReverseEnd,
-      diagonalReverseOffset,
-      forwardPoints,
-      reversePoints,
-      originalIndex,
-      originalEnd,
-      midOriginalArr,
-      modifiedIndex,
-      modifiedEnd,
-      midModifiedArr,
-      deltaIsEven,
-      quitEarlyArr
-    );
-  }
-  /**
-   * Shifts the given changes to provide a more intuitive diff.
-   * While the first element in a diff matches the first element after the diff,
-   * we shift the diff down.
-   *
-   * @param changes The list of changes to shift
-   * @returns The shifted changes
-   */
-  PrettifyChanges(changes) {
-    for (let i = 0; i < changes.length; i++) {
-      const change = changes[i];
-      const originalStop = i < changes.length - 1 ? changes[i + 1].originalStart : this._originalElementsOrHash.length;
-      const modifiedStop = i < changes.length - 1 ? changes[i + 1].modifiedStart : this._modifiedElementsOrHash.length;
-      const checkOriginal = change.originalLength > 0;
-      const checkModified = change.modifiedLength > 0;
-      while (change.originalStart + change.originalLength < originalStop && change.modifiedStart + change.modifiedLength < modifiedStop && (!checkOriginal || this.OriginalElementsAreEqual(change.originalStart, change.originalStart + change.originalLength)) && (!checkModified || this.ModifiedElementsAreEqual(change.modifiedStart, change.modifiedStart + change.modifiedLength))) {
-        const startStrictEqual = this.ElementsAreStrictEqual(change.originalStart, change.modifiedStart);
-        const endStrictEqual = this.ElementsAreStrictEqual(change.originalStart + change.originalLength, change.modifiedStart + change.modifiedLength);
-        if (endStrictEqual && !startStrictEqual) {
-          break;
-        }
-        change.originalStart++;
-        change.modifiedStart++;
-      }
-      const mergedChangeArr = [null];
-      if (i < changes.length - 1 && this.ChangesOverlap(changes[i], changes[i + 1], mergedChangeArr)) {
-        changes[i] = mergedChangeArr[0];
-        changes.splice(i + 1, 1);
-        i--;
-        continue;
-      }
-    }
-    for (let i = changes.length - 1; i >= 0; i--) {
-      const change = changes[i];
-      let originalStop = 0;
-      let modifiedStop = 0;
-      if (i > 0) {
-        const prevChange = changes[i - 1];
-        originalStop = prevChange.originalStart + prevChange.originalLength;
-        modifiedStop = prevChange.modifiedStart + prevChange.modifiedLength;
-      }
-      const checkOriginal = change.originalLength > 0;
-      const checkModified = change.modifiedLength > 0;
-      let bestDelta = 0;
-      let bestScore = this._boundaryScore(change.originalStart, change.originalLength, change.modifiedStart, change.modifiedLength);
-      for (let delta = 1; ; delta++) {
-        const originalStart = change.originalStart - delta;
-        const modifiedStart = change.modifiedStart - delta;
-        if (originalStart < originalStop || modifiedStart < modifiedStop) {
-          break;
-        }
-        if (checkOriginal && !this.OriginalElementsAreEqual(originalStart, originalStart + change.originalLength)) {
-          break;
-        }
-        if (checkModified && !this.ModifiedElementsAreEqual(modifiedStart, modifiedStart + change.modifiedLength)) {
-          break;
-        }
-        const touchingPreviousChange = originalStart === originalStop && modifiedStart === modifiedStop;
-        const score = (touchingPreviousChange ? 5 : 0) + this._boundaryScore(originalStart, change.originalLength, modifiedStart, change.modifiedLength);
-        if (score > bestScore) {
-          bestScore = score;
-          bestDelta = delta;
-        }
-      }
-      change.originalStart -= bestDelta;
-      change.modifiedStart -= bestDelta;
-      const mergedChangeArr = [null];
-      if (i > 0 && this.ChangesOverlap(changes[i - 1], changes[i], mergedChangeArr)) {
-        changes[i - 1] = mergedChangeArr[0];
-        changes.splice(i, 1);
-        i++;
-        continue;
-      }
-    }
-    if (this._hasStrings) {
-      for (let i = 1, len = changes.length; i < len; i++) {
-        const aChange = changes[i - 1];
-        const bChange = changes[i];
-        const matchedLength = bChange.originalStart - aChange.originalStart - aChange.originalLength;
-        const aOriginalStart = aChange.originalStart;
-        const bOriginalEnd = bChange.originalStart + bChange.originalLength;
-        const abOriginalLength = bOriginalEnd - aOriginalStart;
-        const aModifiedStart = aChange.modifiedStart;
-        const bModifiedEnd = bChange.modifiedStart + bChange.modifiedLength;
-        const abModifiedLength = bModifiedEnd - aModifiedStart;
-        if (matchedLength < 5 && abOriginalLength < 20 && abModifiedLength < 20) {
-          const t = this._findBetterContiguousSequence(
-            aOriginalStart,
-            abOriginalLength,
-            aModifiedStart,
-            abModifiedLength,
-            matchedLength
-          );
-          if (t) {
-            const [originalMatchStart, modifiedMatchStart] = t;
-            if (originalMatchStart !== aChange.originalStart + aChange.originalLength || modifiedMatchStart !== aChange.modifiedStart + aChange.modifiedLength) {
-              aChange.originalLength = originalMatchStart - aChange.originalStart;
-              aChange.modifiedLength = modifiedMatchStart - aChange.modifiedStart;
-              bChange.originalStart = originalMatchStart + matchedLength;
-              bChange.modifiedStart = modifiedMatchStart + matchedLength;
-              bChange.originalLength = bOriginalEnd - bChange.originalStart;
-              bChange.modifiedLength = bModifiedEnd - bChange.modifiedStart;
-            }
-          }
-        }
-      }
-    }
-    return changes;
-  }
-  _findBetterContiguousSequence(originalStart, originalLength, modifiedStart, modifiedLength, desiredLength) {
-    if (originalLength < desiredLength || modifiedLength < desiredLength) {
-      return null;
-    }
-    const originalMax = originalStart + originalLength - desiredLength + 1;
-    const modifiedMax = modifiedStart + modifiedLength - desiredLength + 1;
-    let bestScore = 0;
-    let bestOriginalStart = 0;
-    let bestModifiedStart = 0;
-    for (let i = originalStart; i < originalMax; i++) {
-      for (let j = modifiedStart; j < modifiedMax; j++) {
-        const score = this._contiguousSequenceScore(i, j, desiredLength);
-        if (score > 0 && score > bestScore) {
-          bestScore = score;
-          bestOriginalStart = i;
-          bestModifiedStart = j;
-        }
-      }
-    }
-    if (bestScore > 0) {
-      return [bestOriginalStart, bestModifiedStart];
-    }
-    return null;
-  }
-  _contiguousSequenceScore(originalStart, modifiedStart, length) {
-    let score = 0;
-    for (let l = 0; l < length; l++) {
-      if (!this.ElementsAreEqual(originalStart + l, modifiedStart + l)) {
-        return 0;
-      }
-      score += this._originalStringElements[originalStart + l].length;
-    }
-    return score;
-  }
-  _OriginalIsBoundary(index) {
-    if (index <= 0 || index >= this._originalElementsOrHash.length - 1) {
-      return true;
-    }
-    return this._hasStrings && /^\s*$/.test(this._originalStringElements[index]);
-  }
-  _OriginalRegionIsBoundary(originalStart, originalLength) {
-    if (this._OriginalIsBoundary(originalStart) || this._OriginalIsBoundary(originalStart - 1)) {
-      return true;
-    }
-    if (originalLength > 0) {
-      const originalEnd = originalStart + originalLength;
-      if (this._OriginalIsBoundary(originalEnd - 1) || this._OriginalIsBoundary(originalEnd)) {
-        return true;
-      }
-    }
-    return false;
-  }
-  _ModifiedIsBoundary(index) {
-    if (index <= 0 || index >= this._modifiedElementsOrHash.length - 1) {
-      return true;
-    }
-    return this._hasStrings && /^\s*$/.test(this._modifiedStringElements[index]);
-  }
-  _ModifiedRegionIsBoundary(modifiedStart, modifiedLength) {
-    if (this._ModifiedIsBoundary(modifiedStart) || this._ModifiedIsBoundary(modifiedStart - 1)) {
-      return true;
-    }
-    if (modifiedLength > 0) {
-      const modifiedEnd = modifiedStart + modifiedLength;
-      if (this._ModifiedIsBoundary(modifiedEnd - 1) || this._ModifiedIsBoundary(modifiedEnd)) {
-        return true;
-      }
-    }
-    return false;
-  }
-  _boundaryScore(originalStart, originalLength, modifiedStart, modifiedLength) {
-    const originalScore = this._OriginalRegionIsBoundary(originalStart, originalLength) ? 1 : 0;
-    const modifiedScore = this._ModifiedRegionIsBoundary(modifiedStart, modifiedLength) ? 1 : 0;
-    return originalScore + modifiedScore;
-  }
-  /**
-   * Concatenates the two input DiffChange lists and returns the resulting
-   * list.
-   * @param The left changes
-   * @param The right changes
-   * @returns The concatenated list
-   */
-  ConcatenateChanges(left, right) {
-    const mergedChangeArr = [];
-    if (left.length === 0 || right.length === 0) {
-      return right.length > 0 ? right : left;
-    } else if (this.ChangesOverlap(left[left.length - 1], right[0], mergedChangeArr)) {
-      const result = new Array(left.length + right.length - 1);
-      MyArray.Copy(left, 0, result, 0, left.length - 1);
-      result[left.length - 1] = mergedChangeArr[0];
-      MyArray.Copy(right, 1, result, left.length, right.length - 1);
-      return result;
-    } else {
-      const result = new Array(left.length + right.length);
-      MyArray.Copy(left, 0, result, 0, left.length);
-      MyArray.Copy(right, 0, result, left.length, right.length);
-      return result;
-    }
-  }
-  /**
-   * Returns true if the two changes overlap and can be merged into a single
-   * change
-   * @param left The left change
-   * @param right The right change
-   * @param mergedChange The merged change if the two overlap, null otherwise
-   * @returns True if the two changes overlap
-   */
-  ChangesOverlap(left, right, mergedChangeArr) {
-    Debug.Assert(left.originalStart <= right.originalStart, "Left change is not less than or equal to right change");
-    Debug.Assert(left.modifiedStart <= right.modifiedStart, "Left change is not less than or equal to right change");
-    if (left.originalStart + left.originalLength >= right.originalStart || left.modifiedStart + left.modifiedLength >= right.modifiedStart) {
-      const originalStart = left.originalStart;
-      let originalLength = left.originalLength;
-      const modifiedStart = left.modifiedStart;
-      let modifiedLength = left.modifiedLength;
-      if (left.originalStart + left.originalLength >= right.originalStart) {
-        originalLength = right.originalStart + right.originalLength - left.originalStart;
-      }
-      if (left.modifiedStart + left.modifiedLength >= right.modifiedStart) {
-        modifiedLength = right.modifiedStart + right.modifiedLength - left.modifiedStart;
-      }
-      mergedChangeArr[0] = new DiffChange(originalStart, originalLength, modifiedStart, modifiedLength);
-      return true;
-    } else {
-      mergedChangeArr[0] = null;
-      return false;
-    }
-  }
-  /**
-   * Helper method used to clip a diagonal index to the range of valid
-   * diagonals. This also decides whether or not the diagonal index,
-   * if it exceeds the boundary, should be clipped to the boundary or clipped
-   * one inside the boundary depending on the Even/Odd status of the boundary
-   * and numDifferences.
-   * @param diagonal The index of the diagonal to clip.
-   * @param numDifferences The current number of differences being iterated upon.
-   * @param diagonalBaseIndex The base reference diagonal.
-   * @param numDiagonals The total number of diagonals.
-   * @returns The clipped diagonal index.
-   */
-  ClipDiagonalBound(diagonal, numDifferences, diagonalBaseIndex, numDiagonals) {
-    if (diagonal >= 0 && diagonal < numDiagonals) {
-      return diagonal;
-    }
-    const diagonalsBelow = diagonalBaseIndex;
-    const diagonalsAbove = numDiagonals - diagonalBaseIndex - 1;
-    const diffEven = numDifferences % 2 === 0;
-    if (diagonal < 0) {
-      const lowerBoundEven = diagonalsBelow % 2 === 0;
-      return diffEven === lowerBoundEven ? 0 : 1;
-    } else {
-      const upperBoundEven = diagonalsAbove % 2 === 0;
-      return diffEven === upperBoundEven ? numDiagonals - 1 : numDiagonals - 2;
-    }
-  }
-}
-
-class LinesDiff {
-  constructor(changes, moves, hitTimeout) {
-    this.changes = changes;
-    this.moves = moves;
-    this.hitTimeout = hitTimeout;
-  }
-}
-
-class BugIndicatingError extends Error {
-  constructor(message) {
-    super(message || "An unexpected bug occurred.");
-    Object.setPrototypeOf(this, BugIndicatingError.prototype);
-  }
-}
-
-class OffsetRange {
-  constructor(start, endExclusive) {
-    this.start = start;
-    this.endExclusive = endExclusive;
-    if (start > endExclusive) {
-      throw new BugIndicatingError(`Invalid range: ${this.toString()}`);
-    }
-  }
-  static fromTo(start, endExclusive) {
-    return new OffsetRange(start, endExclusive);
-  }
-  static addRange(range, sortedRanges) {
-    let i = 0;
-    while (i < sortedRanges.length && sortedRanges[i].endExclusive < range.start) {
-      i++;
-    }
-    let j = i;
-    while (j < sortedRanges.length && sortedRanges[j].start <= range.endExclusive) {
-      j++;
-    }
-    if (i === j) {
-      sortedRanges.splice(i, 0, range);
-    } else {
-      const start = Math.min(range.start, sortedRanges[i].start);
-      const end = Math.max(range.endExclusive, sortedRanges[j - 1].endExclusive);
-      sortedRanges.splice(i, j - i, new OffsetRange(start, end));
-    }
-  }
-  static tryCreate(start, endExclusive) {
-    if (start > endExclusive) {
-      return undefined;
-    }
-    return new OffsetRange(start, endExclusive);
-  }
-  static ofLength(length) {
-    return new OffsetRange(0, length);
-  }
-  static ofStartAndLength(start, length) {
-    return new OffsetRange(start, start + length);
-  }
-  static emptyAt(offset) {
-    return new OffsetRange(offset, offset);
-  }
-  get isEmpty() {
-    return this.start === this.endExclusive;
-  }
-  delta(offset) {
-    return new OffsetRange(this.start + offset, this.endExclusive + offset);
-  }
-  deltaStart(offset) {
-    return new OffsetRange(this.start + offset, this.endExclusive);
-  }
-  deltaEnd(offset) {
-    return new OffsetRange(this.start, this.endExclusive + offset);
-  }
-  get length() {
-    return this.endExclusive - this.start;
-  }
-  toString() {
-    return `[${this.start}, ${this.endExclusive})`;
-  }
-  equals(other) {
-    return this.start === other.start && this.endExclusive === other.endExclusive;
-  }
-  containsRange(other) {
-    return this.start <= other.start && other.endExclusive <= this.endExclusive;
-  }
-  contains(offset) {
-    return this.start <= offset && offset < this.endExclusive;
-  }
-  /**
-   * for all numbers n: range1.contains(n) or range2.contains(n) => range1.join(range2).contains(n)
-   * The joined range is the smallest range that contains both ranges.
-   */
-  join(other) {
-    return new OffsetRange(Math.min(this.start, other.start), Math.max(this.endExclusive, other.endExclusive));
-  }
-  /**
-   * for all numbers n: range1.contains(n) and range2.contains(n) <=> range1.intersect(range2).contains(n)
-   *
-   * The resulting range is empty if the ranges do not intersect, but touch.
-   * If the ranges don't even touch, the result is undefined.
-   */
-  intersect(other) {
-    const start = Math.max(this.start, other.start);
-    const end = Math.min(this.endExclusive, other.endExclusive);
-    if (start <= end) {
-      return new OffsetRange(start, end);
-    }
-    return undefined;
-  }
-  intersectionLength(range) {
-    const start = Math.max(this.start, range.start);
-    const end = Math.min(this.endExclusive, range.endExclusive);
-    return Math.max(0, end - start);
-  }
-  intersects(other) {
-    const start = Math.max(this.start, other.start);
-    const end = Math.min(this.endExclusive, other.endExclusive);
-    return start < end;
-  }
-  intersectsOrTouches(other) {
-    const start = Math.max(this.start, other.start);
-    const end = Math.min(this.endExclusive, other.endExclusive);
-    return start <= end;
-  }
-  isBefore(other) {
-    return this.endExclusive <= other.start;
-  }
-  isAfter(other) {
-    return this.start >= other.endExclusive;
-  }
-  slice(arr) {
-    return arr.slice(this.start, this.endExclusive);
-  }
-  substring(str) {
-    return str.substring(this.start, this.endExclusive);
-  }
-  /**
-   * Returns the given value if it is contained in this instance, otherwise the closest value that is contained.
-   * The range must not be empty.
-   */
-  clip(value) {
-    if (this.isEmpty) {
-      throw new BugIndicatingError(`Invalid clipping range: ${this.toString()}`);
-    }
-    return Math.max(this.start, Math.min(this.endExclusive - 1, value));
-  }
-  /**
-   * Returns `r := value + k * length` such that `r` is contained in this range.
-   * The range must not be empty.
-   *
-   * E.g. `[5, 10).clipCyclic(10) === 5`, `[5, 10).clipCyclic(11) === 6` and `[5, 10).clipCyclic(4) === 9`.
-   */
-  clipCyclic(value) {
-    if (this.isEmpty) {
-      throw new BugIndicatingError(`Invalid clipping range: ${this.toString()}`);
-    }
-    if (value < this.start) {
-      return this.endExclusive - (this.start - value) % this.length;
-    }
-    if (value >= this.endExclusive) {
-      return this.start + (value - this.start) % this.length;
-    }
-    return value;
-  }
-  map(f) {
-    const result = [];
-    for (let i = this.start; i < this.endExclusive; i++) {
-      result.push(f(i));
-    }
-    return result;
-  }
-  forEach(f) {
-    for (let i = this.start; i < this.endExclusive; i++) {
-      f(i);
-    }
-  }
-}
-
-class Position {
-  constructor(lineNumber, column) {
-    this.lineNumber = lineNumber;
-    this.column = column;
-  }
-  /**
-   * Test if this position equals other position
-   */
-  equals(other) {
-    return Position.equals(this, other);
-  }
-  /**
-   * Test if position `a` equals position `b`
-   */
-  static equals(a, b) {
-    if (!a && !b) {
-      return true;
-    }
-    return !!a && !!b && a.lineNumber === b.lineNumber && a.column === b.column;
-  }
-  /**
-   * Test if this position is before other position.
-   * If the two positions are equal, the result will be false.
-   */
-  isBefore(other) {
-    return Position.isBefore(this, other);
-  }
-  /**
-   * Test if position `a` is before position `b`.
-   * If the two positions are equal, the result will be false.
-   */
-  static isBefore(a, b) {
-    if (a.lineNumber < b.lineNumber) {
-      return true;
-    }
-    if (b.lineNumber < a.lineNumber) {
-      return false;
-    }
-    return a.column < b.column;
-  }
-  /**
-   * Test if this position is before other position.
-   * If the two positions are equal, the result will be true.
-   */
-  isBeforeOrEqual(other) {
-    return Position.isBeforeOrEqual(this, other);
-  }
-  /**
-   * Test if position `a` is before position `b`.
-   * If the two positions are equal, the result will be true.
-   */
-  static isBeforeOrEqual(a, b) {
-    if (a.lineNumber < b.lineNumber) {
-      return true;
-    }
-    if (b.lineNumber < a.lineNumber) {
-      return false;
-    }
-    return a.column <= b.column;
-  }
-}
-
-class Range {
-  constructor(startLineNumber, startColumn, endLineNumber, endColumn) {
-    if (startLineNumber > endLineNumber || startLineNumber === endLineNumber && startColumn > endColumn) {
-      this.startLineNumber = endLineNumber;
-      this.startColumn = endColumn;
-      this.endLineNumber = startLineNumber;
-      this.endColumn = startColumn;
-    } else {
-      this.startLineNumber = startLineNumber;
-      this.startColumn = startColumn;
-      this.endLineNumber = endLineNumber;
-      this.endColumn = endColumn;
-    }
-  }
-  /**
-   * Test if this range is empty.
-   */
-  isEmpty() {
-    return Range.isEmpty(this);
-  }
-  /**
-   * Test if `range` is empty.
-   */
-  static isEmpty(range) {
-    return range.startLineNumber === range.endLineNumber && range.startColumn === range.endColumn;
-  }
-  /**
-   * Test if position is in this range. If the position is at the edges, will return true.
-   */
-  containsPosition(position) {
-    return Range.containsPosition(this, position);
-  }
-  /**
-   * Test if `position` is in `range`. If the position is at the edges, will return true.
-   */
-  static containsPosition(range, position) {
-    if (position.lineNumber < range.startLineNumber || position.lineNumber > range.endLineNumber) {
-      return false;
-    }
-    if (position.lineNumber === range.startLineNumber && position.column < range.startColumn) {
-      return false;
-    }
-    if (position.lineNumber === range.endLineNumber && position.column > range.endColumn) {
-      return false;
-    }
-    return true;
-  }
-  /**
-   * Test if range is in this range. If the range is equal to this range, will return true.
-   */
-  containsRange(range) {
-    return Range.containsRange(this, range);
-  }
-  /**
-   * Test if `otherRange` is in `range`. If the ranges are equal, will return true.
-   */
-  static containsRange(range, otherRange) {
-    if (otherRange.startLineNumber < range.startLineNumber || otherRange.endLineNumber < range.startLineNumber) {
-      return false;
-    }
-    if (otherRange.startLineNumber > range.endLineNumber || otherRange.endLineNumber > range.endLineNumber) {
-      return false;
-    }
-    if (otherRange.startLineNumber === range.startLineNumber && otherRange.startColumn < range.startColumn) {
-      return false;
-    }
-    if (otherRange.endLineNumber === range.endLineNumber && otherRange.endColumn > range.endColumn) {
-      return false;
-    }
-    return true;
-  }
-  /**
-   * Test if `range` is strictly in this range. `range` must start after and end before this range for the result to be true.
-   */
-  strictContainsRange(range) {
-    return Range.strictContainsRange(this, range);
-  }
-  /**
-   * Test if `otherRange` is strictly in `range` (must start after, and end before). If the ranges are equal, will return false.
-   */
-  static strictContainsRange(range, otherRange) {
-    if (otherRange.startLineNumber < range.startLineNumber || otherRange.endLineNumber < range.startLineNumber) {
-      return false;
-    }
-    if (otherRange.startLineNumber > range.endLineNumber || otherRange.endLineNumber > range.endLineNumber) {
-      return false;
-    }
-    if (otherRange.startLineNumber === range.startLineNumber && otherRange.startColumn <= range.startColumn) {
-      return false;
-    }
-    if (otherRange.endLineNumber === range.endLineNumber && otherRange.endColumn >= range.endColumn) {
-      return false;
-    }
-    return true;
-  }
-  /**
-   * A reunion of the two ranges.
-   * The smallest position will be used as the start point, and the largest one as the end point.
-   */
-  plusRange(range) {
-    return Range.plusRange(this, range);
-  }
-  /**
-   * A reunion of the two ranges.
-   * The smallest position will be used as the start point, and the largest one as the end point.
-   */
-  static plusRange(a, b) {
-    let startLineNumber;
-    let startColumn;
-    let endLineNumber;
-    let endColumn;
-    if (b.startLineNumber < a.startLineNumber) {
-      startLineNumber = b.startLineNumber;
-      startColumn = b.startColumn;
-    } else if (b.startLineNumber === a.startLineNumber) {
-      startLineNumber = b.startLineNumber;
-      startColumn = Math.min(b.startColumn, a.startColumn);
-    } else {
-      startLineNumber = a.startLineNumber;
-      startColumn = a.startColumn;
-    }
-    if (b.endLineNumber > a.endLineNumber) {
-      endLineNumber = b.endLineNumber;
-      endColumn = b.endColumn;
-    } else if (b.endLineNumber === a.endLineNumber) {
-      endLineNumber = b.endLineNumber;
-      endColumn = Math.max(b.endColumn, a.endColumn);
-    } else {
-      endLineNumber = a.endLineNumber;
-      endColumn = a.endColumn;
-    }
-    return new Range(startLineNumber, startColumn, endLineNumber, endColumn);
-  }
-  /**
-   * A intersection of the two ranges.
-   */
-  intersectRanges(range) {
-    return Range.intersectRanges(this, range);
-  }
-  /**
-   * A intersection of the two ranges.
-   */
-  static intersectRanges(a, b) {
-    let resultStartLineNumber = a.startLineNumber;
-    let resultStartColumn = a.startColumn;
-    let resultEndLineNumber = a.endLineNumber;
-    let resultEndColumn = a.endColumn;
-    const otherStartLineNumber = b.startLineNumber;
-    const otherStartColumn = b.startColumn;
-    const otherEndLineNumber = b.endLineNumber;
-    const otherEndColumn = b.endColumn;
-    if (resultStartLineNumber < otherStartLineNumber) {
-      resultStartLineNumber = otherStartLineNumber;
-      resultStartColumn = otherStartColumn;
-    } else if (resultStartLineNumber === otherStartLineNumber) {
-      resultStartColumn = Math.max(resultStartColumn, otherStartColumn);
-    }
-    if (resultEndLineNumber > otherEndLineNumber) {
-      resultEndLineNumber = otherEndLineNumber;
-      resultEndColumn = otherEndColumn;
-    } else if (resultEndLineNumber === otherEndLineNumber) {
-      resultEndColumn = Math.min(resultEndColumn, otherEndColumn);
-    }
-    if (resultStartLineNumber > resultEndLineNumber) {
-      return null;
-    }
-    if (resultStartLineNumber === resultEndLineNumber && resultStartColumn > resultEndColumn) {
-      return null;
-    }
-    return new Range(resultStartLineNumber, resultStartColumn, resultEndLineNumber, resultEndColumn);
-  }
-  /**
-   * Test if this range equals other.
-   */
-  equalsRange(other) {
-    return Range.equalsRange(this, other);
-  }
-  /**
-   * Test if range `a` equals `b`.
-   */
-  static equalsRange(a, b) {
-    if (!a && !b) {
-      return true;
-    }
-    return !!a && !!b && a.startLineNumber === b.startLineNumber && a.startColumn === b.startColumn && a.endLineNumber === b.endLineNumber && a.endColumn === b.endColumn;
-  }
-  /**
-   * Return the end position (which will be after or equal to the start position)
-   */
-  getEndPosition() {
-    return Range.getEndPosition(this);
-  }
-  /**
-   * Return the end position (which will be after or equal to the start position)
-   */
-  static getEndPosition(range) {
-    return new Position(range.endLineNumber, range.endColumn);
-  }
-  /**
-   * Return the start position (which will be before or equal to the end position)
-   */
-  getStartPosition() {
-    return Range.getStartPosition(this);
-  }
-  /**
-   * Return the start position (which will be before or equal to the end position)
-   */
-  static getStartPosition(range) {
-    return new Position(range.startLineNumber, range.startColumn);
-  }
-  /**
-   * Create a new empty range using this range's start position.
-   */
-  collapseToStart() {
-    return Range.collapseToStart(this);
-  }
-  /**
-   * Create a new empty range using this range's start position.
-   */
-  static collapseToStart(range) {
-    return new Range(range.startLineNumber, range.startColumn, range.startLineNumber, range.startColumn);
-  }
-  /**
-   * Create a new empty range using this range's end position.
-   */
-  collapseToEnd() {
-    return Range.collapseToEnd(this);
-  }
-  /**
-   * Create a new empty range using this range's end position.
-   */
-  static collapseToEnd(range) {
-    return new Range(range.endLineNumber, range.endColumn, range.endLineNumber, range.endColumn);
-  }
-  // ---
-  static fromPositions(start, end = start) {
-    return new Range(start.lineNumber, start.column, end.lineNumber, end.column);
-  }
-}
-
-class LineRange {
-  static fromRangeInclusive(range) {
-    return new LineRange(range.startLineNumber, range.endLineNumber + 1);
-  }
-  static join(lineRanges) {
-    if (lineRanges.length === 0) {
-      throw new BugIndicatingError("lineRanges cannot be empty");
-    }
-    let startLineNumber = lineRanges[0].startLineNumber;
-    let endLineNumberExclusive = lineRanges[0].endLineNumberExclusive;
-    for (let i = 1; i < lineRanges.length; i++) {
-      startLineNumber = Math.min(startLineNumber, lineRanges[i].startLineNumber);
-      endLineNumberExclusive = Math.max(endLineNumberExclusive, lineRanges[i].endLineNumberExclusive);
-    }
-    return new LineRange(startLineNumber, endLineNumberExclusive);
-  }
-  static ofLength(startLineNumber, length) {
-    return new LineRange(startLineNumber, startLineNumber + length);
-  }
-  constructor(startLineNumber, endLineNumberExclusive) {
-    if (startLineNumber > endLineNumberExclusive) {
-      throw new BugIndicatingError(`startLineNumber ${startLineNumber} cannot be after endLineNumberExclusive ${endLineNumberExclusive}`);
-    }
-    this.startLineNumber = startLineNumber;
-    this.endLineNumberExclusive = endLineNumberExclusive;
-  }
-  /**
-   * Indicates if this line range is empty.
-   */
-  get isEmpty() {
-    return this.startLineNumber === this.endLineNumberExclusive;
-  }
-  /**
-   * Moves this line range by the given offset of line numbers.
-   */
-  delta(offset) {
-    return new LineRange(this.startLineNumber + offset, this.endLineNumberExclusive + offset);
-  }
-  /**
-   * The number of lines this line range spans.
-   */
-  get length() {
-    return this.endLineNumberExclusive - this.startLineNumber;
-  }
-  /**
-   * Creates a line range that combines this and the given line range.
-   */
-  join(other) {
-    return new LineRange(
-      Math.min(this.startLineNumber, other.startLineNumber),
-      Math.max(this.endLineNumberExclusive, other.endLineNumberExclusive)
-    );
-  }
-  /**
-   * The resulting range is empty if the ranges do not intersect, but touch.
-   * If the ranges don't even touch, the result is undefined.
-   */
-  intersect(other) {
-    const startLineNumber = Math.max(this.startLineNumber, other.startLineNumber);
-    const endLineNumberExclusive = Math.min(this.endLineNumberExclusive, other.endLineNumberExclusive);
-    if (startLineNumber <= endLineNumberExclusive) {
-      return new LineRange(startLineNumber, endLineNumberExclusive);
-    }
-    return undefined;
-  }
-  overlapOrTouch(other) {
-    return this.startLineNumber <= other.endLineNumberExclusive && other.startLineNumber <= this.endLineNumberExclusive;
-  }
-  toInclusiveRange() {
-    if (this.isEmpty) {
-      return null;
-    }
-    return new Range(this.startLineNumber, 1, this.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER);
-  }
-  /**
-   * Converts this 1-based line range to a 0-based offset range (subtracts 1!).
-   * @internal
-   */
-  toOffsetRange() {
-    return new OffsetRange(this.startLineNumber - 1, this.endLineNumberExclusive - 1);
-  }
-}
-
-class LineRangeMapping {
-  constructor(originalRange, modifiedRange) {
-    this.original = originalRange;
-    this.modified = modifiedRange;
-  }
-  join(other) {
-    return new LineRangeMapping(
-      this.original.join(other.original),
-      this.modified.join(other.modified)
-    );
-  }
-  get changedLineCount() {
-    return Math.max(this.original.length, this.modified.length);
-  }
-  /**
-   * This method assumes that the LineRangeMapping describes a valid diff!
-   * I.e. if one range is empty, the other range cannot be the entire document.
-   * It avoids various problems when the line range points to non-existing line-numbers.
-  */
-  toRangeMapping() {
-    const origInclusiveRange = this.original.toInclusiveRange();
-    const modInclusiveRange = this.modified.toInclusiveRange();
-    if (origInclusiveRange && modInclusiveRange) {
-      return new RangeMapping(origInclusiveRange, modInclusiveRange);
-    } else if (this.original.startLineNumber === 1 || this.modified.startLineNumber === 1) {
-      if (!(this.modified.startLineNumber === 1 && this.original.startLineNumber === 1)) {
-        throw new BugIndicatingError("not a valid diff");
-      }
-      return new RangeMapping(
-        new Range(this.original.startLineNumber, 1, this.original.endLineNumberExclusive, 1),
-        new Range(this.modified.startLineNumber, 1, this.modified.endLineNumberExclusive, 1)
-      );
-    } else {
-      return new RangeMapping(
-        new Range(this.original.startLineNumber - 1, Number.MAX_SAFE_INTEGER, this.original.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER),
-        new Range(this.modified.startLineNumber - 1, Number.MAX_SAFE_INTEGER, this.modified.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER)
-      );
-    }
-  }
-  /**
-   * This method assumes that the LineRangeMapping describes a valid diff!
-   * I.e. if one range is empty, the other range cannot be the entire document.
-   * It avoids various problems when the line range points to non-existing line-numbers.
-  */
-  toRangeMapping2(original, modified) {
-    if (isValidLineNumber(this.original.endLineNumberExclusive, original) && isValidLineNumber(this.modified.endLineNumberExclusive, modified)) {
-      return new RangeMapping(
-        new Range(this.original.startLineNumber, 1, this.original.endLineNumberExclusive, 1),
-        new Range(this.modified.startLineNumber, 1, this.modified.endLineNumberExclusive, 1)
-      );
-    }
-    if (!this.original.isEmpty && !this.modified.isEmpty) {
-      return new RangeMapping(
-        Range.fromPositions(
-          new Position(this.original.startLineNumber, 1),
-          normalizePosition(new Position(this.original.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER), original)
-        ),
-        Range.fromPositions(
-          new Position(this.modified.startLineNumber, 1),
-          normalizePosition(new Position(this.modified.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER), modified)
-        )
-      );
-    }
-    if (this.original.startLineNumber > 1 && this.modified.startLineNumber > 1) {
-      return new RangeMapping(
-        Range.fromPositions(
-          normalizePosition(new Position(this.original.startLineNumber - 1, Number.MAX_SAFE_INTEGER), original),
-          normalizePosition(new Position(this.original.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER), original)
-        ),
-        Range.fromPositions(
-          normalizePosition(new Position(this.modified.startLineNumber - 1, Number.MAX_SAFE_INTEGER), modified),
-          normalizePosition(new Position(this.modified.endLineNumberExclusive - 1, Number.MAX_SAFE_INTEGER), modified)
-        )
-      );
-    }
-    throw new BugIndicatingError();
-  }
-}
-function normalizePosition(position, content) {
-  if (position.lineNumber < 1) {
-    return new Position(1, 1);
-  }
-  if (position.lineNumber > content.length) {
-    return new Position(content.length, content[content.length - 1].length + 1);
-  }
-  const line = content[position.lineNumber - 1];
-  if (position.column > line.length + 1) {
-    return new Position(position.lineNumber, line.length + 1);
-  }
-  return position;
-}
-function isValidLineNumber(lineNumber, lines) {
-  return lineNumber >= 1 && lineNumber <= lines.length;
-}
-class DetailedLineRangeMapping extends LineRangeMapping {
-  static fromRangeMappings(rangeMappings) {
-    const originalRange = LineRange.join(rangeMappings.map((r) => LineRange.fromRangeInclusive(r.originalRange)));
-    const modifiedRange = LineRange.join(rangeMappings.map((r) => LineRange.fromRangeInclusive(r.modifiedRange)));
-    return new DetailedLineRangeMapping(originalRange, modifiedRange, rangeMappings);
-  }
-  constructor(originalRange, modifiedRange, innerChanges) {
-    super(originalRange, modifiedRange);
-    this.innerChanges = innerChanges;
-  }
-  flip() {
-    return new DetailedLineRangeMapping(this.modified, this.original, this.innerChanges?.map((c) => c.flip()));
-  }
-  withInnerChangesFromLineRanges() {
-    return new DetailedLineRangeMapping(this.original, this.modified, [this.toRangeMapping()]);
-  }
-}
-class RangeMapping {
-  static join(rangeMappings) {
-    if (rangeMappings.length === 0) {
-      throw new BugIndicatingError("Cannot join an empty list of range mappings");
-    }
-    let result = rangeMappings[0];
-    for (let i = 1; i < rangeMappings.length; i++) {
-      result = result.join(rangeMappings[i]);
-    }
-    return result;
-  }
-  static assertSorted(rangeMappings) {
-    for (let i = 1; i < rangeMappings.length; i++) {
-      const previous = rangeMappings[i - 1];
-      const current = rangeMappings[i];
-      if (!(previous.originalRange.getEndPosition().isBeforeOrEqual(current.originalRange.getStartPosition()) && previous.modifiedRange.getEndPosition().isBeforeOrEqual(current.modifiedRange.getStartPosition()))) {
-        throw new BugIndicatingError("Range mappings must be sorted");
-      }
-    }
-  }
-  constructor(originalRange, modifiedRange) {
-    this.originalRange = originalRange;
-    this.modifiedRange = modifiedRange;
-  }
-  flip() {
-    return new RangeMapping(this.modifiedRange, this.originalRange);
-  }
-  /**
-   * Creates a single text edit that describes the change from the original to the modified text.
-  */
-  join(other) {
-    return new RangeMapping(
-      this.originalRange.plusRange(other.originalRange),
-      this.modifiedRange.plusRange(other.modifiedRange)
-    );
-  }
-}
-
-const MINIMUM_MATCHING_CHARACTER_LENGTH = 3;
-class LegacyLinesDiffComputer {
-  computeDiff(originalLines, modifiedLines, options) {
-    const diffComputer = new DiffComputer(originalLines, modifiedLines, {
-      maxComputationTime: options.maxComputationTimeMs,
-      shouldIgnoreTrimWhitespace: options.ignoreTrimWhitespace,
-      shouldComputeCharChanges: true,
-      shouldMakePrettyDiff: true,
-      shouldPostProcessCharChanges: true
-    });
-    const result = diffComputer.computeDiff();
-    const changes = [];
-    let lastChange = null;
-    for (const c of result.changes) {
-      let originalRange;
-      if (c.originalEndLineNumber === 0) {
-        originalRange = new LineRange(c.originalStartLineNumber + 1, c.originalStartLineNumber + 1);
-      } else {
-        originalRange = new LineRange(c.originalStartLineNumber, c.originalEndLineNumber + 1);
-      }
-      let modifiedRange;
-      if (c.modifiedEndLineNumber === 0) {
-        modifiedRange = new LineRange(c.modifiedStartLineNumber + 1, c.modifiedStartLineNumber + 1);
-      } else {
-        modifiedRange = new LineRange(c.modifiedStartLineNumber, c.modifiedEndLineNumber + 1);
-      }
-      let change = new DetailedLineRangeMapping(originalRange, modifiedRange, c.charChanges?.map((c2) => new RangeMapping(
-        new Range(c2.originalStartLineNumber, c2.originalStartColumn, c2.originalEndLineNumber, c2.originalEndColumn),
-        new Range(c2.modifiedStartLineNumber, c2.modifiedStartColumn, c2.modifiedEndLineNumber, c2.modifiedEndColumn)
-      )));
-      if (lastChange) {
-        if (lastChange.modified.endLineNumberExclusive === change.modified.startLineNumber || lastChange.original.endLineNumberExclusive === change.original.startLineNumber) {
-          change = new DetailedLineRangeMapping(
-            lastChange.original.join(change.original),
-            lastChange.modified.join(change.modified),
-            lastChange.innerChanges && change.innerChanges ? lastChange.innerChanges.concat(change.innerChanges) : undefined
-          );
-          changes.pop();
-        }
-      }
-      changes.push(change);
-      lastChange = change;
-    }
-    return new LinesDiff(changes, [], result.quitEarly);
-  }
-}
-function computeDiff$1(originalSequence, modifiedSequence, continueProcessingPredicate, pretty) {
-  const diffAlgo = new LcsDiff(originalSequence, modifiedSequence, continueProcessingPredicate);
-  return diffAlgo.ComputeDiff(pretty);
-}
-class LineSequence {
-  constructor(lines) {
-    const startColumns = [];
-    const endColumns = [];
-    for (let i = 0, length = lines.length; i < length; i++) {
-      startColumns[i] = getFirstNonBlankColumn(lines[i], 1);
-      endColumns[i] = getLastNonBlankColumn(lines[i], 1);
-    }
-    this.lines = lines;
-    this._startColumns = startColumns;
-    this._endColumns = endColumns;
-  }
-  getElements() {
-    const elements = [];
-    for (let i = 0, len = this.lines.length; i < len; i++) {
-      elements[i] = this.lines[i].substring(this._startColumns[i] - 1, this._endColumns[i] - 1);
-    }
-    return elements;
-  }
-  getStrictElement(index) {
-    return this.lines[index];
-  }
-  getStartLineNumber(i) {
-    return i + 1;
-  }
-  getEndLineNumber(i) {
-    return i + 1;
-  }
-  createCharSequence(shouldIgnoreTrimWhitespace, startIndex, endIndex) {
-    const charCodes = [];
-    const lineNumbers = [];
-    const columns = [];
-    let len = 0;
-    for (let index = startIndex; index <= endIndex; index++) {
-      const lineContent = this.lines[index];
-      const startColumn = shouldIgnoreTrimWhitespace ? this._startColumns[index] : 1;
-      const endColumn = shouldIgnoreTrimWhitespace ? this._endColumns[index] : lineContent.length + 1;
-      for (let col = startColumn; col < endColumn; col++) {
-        charCodes[len] = lineContent.charCodeAt(col - 1);
-        lineNumbers[len] = index + 1;
-        columns[len] = col;
-        len++;
-      }
-      if (!shouldIgnoreTrimWhitespace && index < endIndex) {
-        charCodes[len] = 10;
-        lineNumbers[len] = index + 1;
-        columns[len] = lineContent.length + 1;
-        len++;
-      }
-    }
-    return new CharSequence(charCodes, lineNumbers, columns);
-  }
-}
-class CharSequence {
-  constructor(charCodes, lineNumbers, columns) {
-    this._charCodes = charCodes;
-    this._lineNumbers = lineNumbers;
-    this._columns = columns;
-  }
-  toString() {
-    return "[" + this._charCodes.map((s, idx) => (s === 10 ? "\\n" : String.fromCharCode(s)) + `-(${this._lineNumbers[idx]},${this._columns[idx]})`).join(", ") + "]";
-  }
-  _assertIndex(index, arr) {
-    if (index < 0 || index >= arr.length) {
-      throw new Error(`Illegal index`);
-    }
-  }
-  getElements() {
-    return this._charCodes;
-  }
-  getStartLineNumber(i) {
-    if (i > 0 && i === this._lineNumbers.length) {
-      return this.getEndLineNumber(i - 1);
-    }
-    this._assertIndex(i, this._lineNumbers);
-    return this._lineNumbers[i];
-  }
-  getEndLineNumber(i) {
-    if (i === -1) {
-      return this.getStartLineNumber(i + 1);
-    }
-    this._assertIndex(i, this._lineNumbers);
-    if (this._charCodes[i] === 10) {
-      return this._lineNumbers[i] + 1;
-    }
-    return this._lineNumbers[i];
-  }
-  getStartColumn(i) {
-    if (i > 0 && i === this._columns.length) {
-      return this.getEndColumn(i - 1);
-    }
-    this._assertIndex(i, this._columns);
-    return this._columns[i];
-  }
-  getEndColumn(i) {
-    if (i === -1) {
-      return this.getStartColumn(i + 1);
-    }
-    this._assertIndex(i, this._columns);
-    if (this._charCodes[i] === 10) {
-      return 1;
-    }
-    return this._columns[i] + 1;
-  }
-}
-class CharChange {
-  constructor(originalStartLineNumber, originalStartColumn, originalEndLineNumber, originalEndColumn, modifiedStartLineNumber, modifiedStartColumn, modifiedEndLineNumber, modifiedEndColumn) {
-    this.originalStartLineNumber = originalStartLineNumber;
-    this.originalStartColumn = originalStartColumn;
-    this.originalEndLineNumber = originalEndLineNumber;
-    this.originalEndColumn = originalEndColumn;
-    this.modifiedStartLineNumber = modifiedStartLineNumber;
-    this.modifiedStartColumn = modifiedStartColumn;
-    this.modifiedEndLineNumber = modifiedEndLineNumber;
-    this.modifiedEndColumn = modifiedEndColumn;
-  }
-  static createFromDiffChange(diffChange, originalCharSequence, modifiedCharSequence) {
-    const originalStartLineNumber = originalCharSequence.getStartLineNumber(diffChange.originalStart);
-    const originalStartColumn = originalCharSequence.getStartColumn(diffChange.originalStart);
-    const originalEndLineNumber = originalCharSequence.getEndLineNumber(diffChange.originalStart + diffChange.originalLength - 1);
-    const originalEndColumn = originalCharSequence.getEndColumn(diffChange.originalStart + diffChange.originalLength - 1);
-    const modifiedStartLineNumber = modifiedCharSequence.getStartLineNumber(diffChange.modifiedStart);
-    const modifiedStartColumn = modifiedCharSequence.getStartColumn(diffChange.modifiedStart);
-    const modifiedEndLineNumber = modifiedCharSequence.getEndLineNumber(diffChange.modifiedStart + diffChange.modifiedLength - 1);
-    const modifiedEndColumn = modifiedCharSequence.getEndColumn(diffChange.modifiedStart + diffChange.modifiedLength - 1);
-    return new CharChange(
-      originalStartLineNumber,
-      originalStartColumn,
-      originalEndLineNumber,
-      originalEndColumn,
-      modifiedStartLineNumber,
-      modifiedStartColumn,
-      modifiedEndLineNumber,
-      modifiedEndColumn
-    );
-  }
-}
-function postProcessCharChanges(rawChanges) {
-  if (rawChanges.length <= 1) {
-    return rawChanges;
-  }
-  const result = [rawChanges[0]];
-  let prevChange = result[0];
-  for (let i = 1, len = rawChanges.length; i < len; i++) {
-    const currChange = rawChanges[i];
-    const originalMatchingLength = currChange.originalStart - (prevChange.originalStart + prevChange.originalLength);
-    const modifiedMatchingLength = currChange.modifiedStart - (prevChange.modifiedStart + prevChange.modifiedLength);
-    const matchingLength = Math.min(originalMatchingLength, modifiedMatchingLength);
-    if (matchingLength < MINIMUM_MATCHING_CHARACTER_LENGTH) {
-      prevChange.originalLength = currChange.originalStart + currChange.originalLength - prevChange.originalStart;
-      prevChange.modifiedLength = currChange.modifiedStart + currChange.modifiedLength - prevChange.modifiedStart;
-    } else {
-      result.push(currChange);
-      prevChange = currChange;
-    }
-  }
-  return result;
-}
-class LineChange {
-  constructor(originalStartLineNumber, originalEndLineNumber, modifiedStartLineNumber, modifiedEndLineNumber, charChanges) {
-    this.originalStartLineNumber = originalStartLineNumber;
-    this.originalEndLineNumber = originalEndLineNumber;
-    this.modifiedStartLineNumber = modifiedStartLineNumber;
-    this.modifiedEndLineNumber = modifiedEndLineNumber;
-    this.charChanges = charChanges;
-  }
-  static createFromDiffResult(shouldIgnoreTrimWhitespace, diffChange, originalLineSequence, modifiedLineSequence, continueCharDiff, shouldComputeCharChanges, shouldPostProcessCharChanges) {
-    let originalStartLineNumber;
-    let originalEndLineNumber;
-    let modifiedStartLineNumber;
-    let modifiedEndLineNumber;
-    let charChanges = undefined;
-    if (diffChange.originalLength === 0) {
-      originalStartLineNumber = originalLineSequence.getStartLineNumber(diffChange.originalStart) - 1;
-      originalEndLineNumber = 0;
-    } else {
-      originalStartLineNumber = originalLineSequence.getStartLineNumber(diffChange.originalStart);
-      originalEndLineNumber = originalLineSequence.getEndLineNumber(diffChange.originalStart + diffChange.originalLength - 1);
-    }
-    if (diffChange.modifiedLength === 0) {
-      modifiedStartLineNumber = modifiedLineSequence.getStartLineNumber(diffChange.modifiedStart) - 1;
-      modifiedEndLineNumber = 0;
-    } else {
-      modifiedStartLineNumber = modifiedLineSequence.getStartLineNumber(diffChange.modifiedStart);
-      modifiedEndLineNumber = modifiedLineSequence.getEndLineNumber(diffChange.modifiedStart + diffChange.modifiedLength - 1);
-    }
-    if (shouldComputeCharChanges && diffChange.originalLength > 0 && diffChange.originalLength < 20 && diffChange.modifiedLength > 0 && diffChange.modifiedLength < 20 && continueCharDiff()) {
-      const originalCharSequence = originalLineSequence.createCharSequence(shouldIgnoreTrimWhitespace, diffChange.originalStart, diffChange.originalStart + diffChange.originalLength - 1);
-      const modifiedCharSequence = modifiedLineSequence.createCharSequence(shouldIgnoreTrimWhitespace, diffChange.modifiedStart, diffChange.modifiedStart + diffChange.modifiedLength - 1);
-      if (originalCharSequence.getElements().length > 0 && modifiedCharSequence.getElements().length > 0) {
-        let rawChanges = computeDiff$1(originalCharSequence, modifiedCharSequence, continueCharDiff, true).changes;
-        if (shouldPostProcessCharChanges) {
-          rawChanges = postProcessCharChanges(rawChanges);
-        }
-        charChanges = [];
-        for (let i = 0, length = rawChanges.length; i < length; i++) {
-          charChanges.push(CharChange.createFromDiffChange(rawChanges[i], originalCharSequence, modifiedCharSequence));
-        }
-      }
-    }
-    return new LineChange(originalStartLineNumber, originalEndLineNumber, modifiedStartLineNumber, modifiedEndLineNumber, charChanges);
-  }
-}
-class DiffComputer {
-  constructor(originalLines, modifiedLines, opts) {
-    this.shouldComputeCharChanges = opts.shouldComputeCharChanges;
-    this.shouldPostProcessCharChanges = opts.shouldPostProcessCharChanges;
-    this.shouldIgnoreTrimWhitespace = opts.shouldIgnoreTrimWhitespace;
-    this.shouldMakePrettyDiff = opts.shouldMakePrettyDiff;
-    this.originalLines = originalLines;
-    this.modifiedLines = modifiedLines;
-    this.original = new LineSequence(originalLines);
-    this.modified = new LineSequence(modifiedLines);
-    this.continueLineDiff = createContinueProcessingPredicate(opts.maxComputationTime);
-    this.continueCharDiff = createContinueProcessingPredicate(opts.maxComputationTime === 0 ? 0 : Math.min(opts.maxComputationTime, 5e3));
-  }
-  computeDiff() {
-    if (this.original.lines.length === 1 && this.original.lines[0].length === 0) {
-      if (this.modified.lines.length === 1 && this.modified.lines[0].length === 0) {
-        return {
-          quitEarly: false,
-          changes: []
-        };
-      }
-      return {
-        quitEarly: false,
-        changes: [{
-          originalStartLineNumber: 1,
-          originalEndLineNumber: 1,
-          modifiedStartLineNumber: 1,
-          modifiedEndLineNumber: this.modified.lines.length,
-          charChanges: undefined
-        }]
-      };
-    }
-    if (this.modified.lines.length === 1 && this.modified.lines[0].length === 0) {
-      return {
-        quitEarly: false,
-        changes: [{
-          originalStartLineNumber: 1,
-          originalEndLineNumber: this.original.lines.length,
-          modifiedStartLineNumber: 1,
-          modifiedEndLineNumber: 1,
-          charChanges: undefined
-        }]
-      };
-    }
-    const diffResult = computeDiff$1(this.original, this.modified, this.continueLineDiff, this.shouldMakePrettyDiff);
-    const rawChanges = diffResult.changes;
-    const quitEarly = diffResult.quitEarly;
-    if (this.shouldIgnoreTrimWhitespace) {
-      const lineChanges = [];
-      for (let i = 0, length = rawChanges.length; i < length; i++) {
-        lineChanges.push(LineChange.createFromDiffResult(this.shouldIgnoreTrimWhitespace, rawChanges[i], this.original, this.modified, this.continueCharDiff, this.shouldComputeCharChanges, this.shouldPostProcessCharChanges));
-      }
-      return {
-        quitEarly,
-        changes: lineChanges
-      };
-    }
-    const result = [];
-    let originalLineIndex = 0;
-    let modifiedLineIndex = 0;
-    for (let i = -1, len = rawChanges.length; i < len; i++) {
-      const nextChange = i + 1 < len ? rawChanges[i + 1] : null;
-      const originalStop = nextChange ? nextChange.originalStart : this.originalLines.length;
-      const modifiedStop = nextChange ? nextChange.modifiedStart : this.modifiedLines.length;
-      while (originalLineIndex < originalStop && modifiedLineIndex < modifiedStop) {
-        const originalLine = this.originalLines[originalLineIndex];
-        const modifiedLine = this.modifiedLines[modifiedLineIndex];
-        if (originalLine !== modifiedLine) {
-          {
-            let originalStartColumn = getFirstNonBlankColumn(originalLine, 1);
-            let modifiedStartColumn = getFirstNonBlankColumn(modifiedLine, 1);
-            while (originalStartColumn > 1 && modifiedStartColumn > 1) {
-              const originalChar = originalLine.charCodeAt(originalStartColumn - 2);
-              const modifiedChar = modifiedLine.charCodeAt(modifiedStartColumn - 2);
-              if (originalChar !== modifiedChar) {
-                break;
-              }
-              originalStartColumn--;
-              modifiedStartColumn--;
-            }
-            if (originalStartColumn > 1 || modifiedStartColumn > 1) {
-              this._pushTrimWhitespaceCharChange(
-                result,
-                originalLineIndex + 1,
-                1,
-                originalStartColumn,
-                modifiedLineIndex + 1,
-                1,
-                modifiedStartColumn
-              );
-            }
-          }
-          {
-            let originalEndColumn = getLastNonBlankColumn(originalLine, 1);
-            let modifiedEndColumn = getLastNonBlankColumn(modifiedLine, 1);
-            const originalMaxColumn = originalLine.length + 1;
-            const modifiedMaxColumn = modifiedLine.length + 1;
-            while (originalEndColumn < originalMaxColumn && modifiedEndColumn < modifiedMaxColumn) {
-              const originalChar = originalLine.charCodeAt(originalEndColumn - 1);
-              const modifiedChar = originalLine.charCodeAt(modifiedEndColumn - 1);
-              if (originalChar !== modifiedChar) {
-                break;
-              }
-              originalEndColumn++;
-              modifiedEndColumn++;
-            }
-            if (originalEndColumn < originalMaxColumn || modifiedEndColumn < modifiedMaxColumn) {
-              this._pushTrimWhitespaceCharChange(
-                result,
-                originalLineIndex + 1,
-                originalEndColumn,
-                originalMaxColumn,
-                modifiedLineIndex + 1,
-                modifiedEndColumn,
-                modifiedMaxColumn
-              );
-            }
-          }
-        }
-        originalLineIndex++;
-        modifiedLineIndex++;
-      }
-      if (nextChange) {
-        result.push(LineChange.createFromDiffResult(this.shouldIgnoreTrimWhitespace, nextChange, this.original, this.modified, this.continueCharDiff, this.shouldComputeCharChanges, this.shouldPostProcessCharChanges));
-        originalLineIndex += nextChange.originalLength;
-        modifiedLineIndex += nextChange.modifiedLength;
-      }
-    }
-    return {
-      quitEarly,
-      changes: result
-    };
-  }
-  _pushTrimWhitespaceCharChange(result, originalLineNumber, originalStartColumn, originalEndColumn, modifiedLineNumber, modifiedStartColumn, modifiedEndColumn) {
-    if (this._mergeTrimWhitespaceCharChange(result, originalLineNumber, originalStartColumn, originalEndColumn, modifiedLineNumber, modifiedStartColumn, modifiedEndColumn)) {
-      return;
-    }
-    let charChanges = undefined;
-    if (this.shouldComputeCharChanges) {
-      charChanges = [new CharChange(
-        originalLineNumber,
-        originalStartColumn,
-        originalLineNumber,
-        originalEndColumn,
-        modifiedLineNumber,
-        modifiedStartColumn,
-        modifiedLineNumber,
-        modifiedEndColumn
-      )];
-    }
-    result.push(new LineChange(
-      originalLineNumber,
-      originalLineNumber,
-      modifiedLineNumber,
-      modifiedLineNumber,
-      charChanges
-    ));
-  }
-  _mergeTrimWhitespaceCharChange(result, originalLineNumber, originalStartColumn, originalEndColumn, modifiedLineNumber, modifiedStartColumn, modifiedEndColumn) {
-    const len = result.length;
-    if (len === 0) {
-      return false;
-    }
-    const prevChange = result[len - 1];
-    if (prevChange.originalEndLineNumber === 0 || prevChange.modifiedEndLineNumber === 0) {
-      return false;
-    }
-    if (prevChange.originalEndLineNumber === originalLineNumber && prevChange.modifiedEndLineNumber === modifiedLineNumber) {
-      if (this.shouldComputeCharChanges && prevChange.charChanges) {
-        prevChange.charChanges.push(new CharChange(
-          originalLineNumber,
-          originalStartColumn,
-          originalLineNumber,
-          originalEndColumn,
-          modifiedLineNumber,
-          modifiedStartColumn,
-          modifiedLineNumber,
-          modifiedEndColumn
-        ));
-      }
-      return true;
-    }
-    if (prevChange.originalEndLineNumber + 1 === originalLineNumber && prevChange.modifiedEndLineNumber + 1 === modifiedLineNumber) {
-      prevChange.originalEndLineNumber = originalLineNumber;
-      prevChange.modifiedEndLineNumber = modifiedLineNumber;
-      if (this.shouldComputeCharChanges && prevChange.charChanges) {
-        prevChange.charChanges.push(new CharChange(
-          originalLineNumber,
-          originalStartColumn,
-          originalLineNumber,
-          originalEndColumn,
-          modifiedLineNumber,
-          modifiedStartColumn,
-          modifiedLineNumber,
-          modifiedEndColumn
-        ));
-      }
-      return true;
-    }
-    return false;
-  }
-}
-function getFirstNonBlankColumn(txt, defaultValue) {
-  const r = firstNonWhitespaceIndex(txt);
-  if (r === -1) {
-    return defaultValue;
-  }
-  return r + 1;
-}
-function getLastNonBlankColumn(txt, defaultValue) {
-  const r = lastNonWhitespaceIndex(txt);
-  if (r === -1) {
-    return defaultValue;
-  }
-  return r + 2;
-}
-function createContinueProcessingPredicate(maximumRuntime) {
-  if (maximumRuntime === 0) {
-    return () => true;
-  }
-  const startTime = Date.now();
-  return () => {
-    return Date.now() - startTime < maximumRuntime;
-  };
-}
-
-function computeDiff(originalLines, modifiedLines, options) {
-  let diffComputer = new LegacyLinesDiffComputer();
-  var result = diffComputer.computeDiff(originalLines, modifiedLines, options);
-  return result?.changes.map((changes) => {
-    let originalStartLineNumber;
-    let originalEndLineNumber;
-    let modifiedStartLineNumber;
-    let modifiedEndLineNumber;
-    let innerChanges = changes.innerChanges;
-    originalStartLineNumber = changes.original.startLineNumber - 1;
-    originalEndLineNumber = changes.original.endLineNumberExclusive - 1;
-    modifiedStartLineNumber = changes.modified.startLineNumber - 1;
-    modifiedEndLineNumber = changes.modified.endLineNumberExclusive - 1;
-    return {
-      origStart: originalStartLineNumber,
-      origEnd: originalEndLineNumber,
-      editStart: modifiedStartLineNumber,
-      editEnd: modifiedEndLineNumber,
-      charChanges: innerChanges?.map((m) => ({
-        originalStartLineNumber: m.originalRange.startLineNumber - 1,
-        originalStartColumn: m.originalRange.startColumn - 1,
-        originalEndLineNumber: m.originalRange.endLineNumber - 1,
-        originalEndColumn: m.originalRange.endColumn - 1,
-        modifiedStartLineNumber: m.modifiedRange.startLineNumber - 1,
-        modifiedStartColumn: m.modifiedRange.startColumn - 1,
-        modifiedEndLineNumber: m.modifiedRange.endLineNumber - 1,
-        modifiedEndColumn: m.modifiedRange.endColumn - 1
-      }))
-    };
-  });
-}
-
-exports.computeDiff = computeDiff;
